requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
billiard==4.2.1
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
certifi==2020.6.20
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-init==25.1.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
flask-talisman==1.0.0
Flask-WTF==1.2.1
greenlet==3.1.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.10
importlib-metadata==6.8.0
importlib_resources==6.4.5
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kombu==5.4.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lml==0.1.0
MarkupSafe==2.1.5
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
passlib==1.7.4
prompt_toolkit==3.0.48
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==5.4.1
redis==5.2.0
requests==2.25.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
SQLAlchemy==2.0.36
systemd-python==234
texttable==1.7.0
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.4
WTForms==3.1.2
zipp==3.17.0
zope.interface==5.4.0
